Add the first vagrant file
This is a starting place.  Bento box project builds these VM's in a
known way, which means if we want to use a different base we can build
out our own fork of their [repo](https://github.com/chef/bento).

Why start with such a minimal image? We want to codify first what is
going into the box. We can optimize the build time later when we stop
making so many changes.

Authored-by: Jim Doty <jdoty@pivotal.io>