rethink Pivotal Workstation Setup starting point
The first attempt at this was to convert the workstation setup
scripts, which were targeted at macOS, to linux.  After trying this
setup out, I have become disallusioned with the kitchen sink approach to
installing software. It sets up the situation where the developer
doesn't know the tools on their machine.  Therefore, the next few
commits will roll that attempt back, and then we will start from a bare
system, and build up as we need tooling.

Authored-by: Jim Doty <jdoty@pivotal.io>